# Conflicts:
#	skywalking-ui/vue.config.js

+++ oap-server/server-starter-boot/server-starter-boot-core/src/main/java/com/piesat/skywalking/vo/CpuVo.java
@@ -3,10 +3,12 @@ package com.piesat.skywalking.vo;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class CpuVo {
     @ApiModelProperty(value = "时间")
     private String timestamp;
     @ApiModelProperty(value = "cpu使用率")
-    private double usage;
+    private BigDecimal usage;
 }
